+++ Arknights/__init__.py
@@ -1,2 +1,2 @@
 # from Arknights.ArknightsShell import ArknightsShell
-# from Arknights.base import ArknightsHelper
+# from Arknights.helper import ArknightsHelper
